requirements.txt
# MinutesAI 필수 패키지

# STT (Speech-to-Text)
openai-whisper>=20231117

# UI
gradio>=4.0.0

# 환경 변수 관리
python-dotenv>=1.0.0

# OpenAI API (LLM 기능용)
openai>=1.0.0

# 기타 의존성
numpy>=1.24.0
torch>=2.0.0
torchaudio>=2.0.0
